// src/redux/models/JobModel.tsx
interface JobModel {
  id: number;
  tenLoaiCongViec: string;
  dsNhomChiTietLoai: DsNhomChiTietLoai[];
}

export interface DsNhomChiTietLoai {
  id: number;
  tenNhom: string;
  hinhAnh: string;
  maLoaiCongviec: number;
  dsChiTietLoai: DsChiTietLoai[];
}

export interface DsChiTietLoai {
  id: number;
  tenChiTiet: string;
}
export interface LoaiCongViec {
  id: number;
  tenLoaiCongViec: string;
}

export interface ThueCongViecViewModel {
  id: number;
  ngayThue: string;
  hoanThanh: boolean;
  congViec: CongViecViewModel;
}

export interface ThueCongViec {
  id: number;
  maCongViec: number;
  maNguoiThue: number;
  ngayThue: string;
  hoanThanh: boolean;
}

export interface CongViecViewModel {
  id: number;
  tenCongViec: string;
  danhGia: number;
  giaTien: number;
  nguoiTao: number;
  hinhAnh: string;
  moTa: string;
  maChiTietLoaiCongViec: number;
  moTaNgan: string;
  saoCongViec: number;
}

export interface ThemCongViecViewModel {
  tenCongViec: string;
  danhGia: number;
  giaTien: number;
  hinhAnh: string;
  moTa: string;
  maChiTietLoaiCongViec: number;
  moTaNgan: string;
  saoCongViec: number;
}

export interface CongViecChiTiet {
  id: number;
  congViec: CongViecViewModel;
  tenLoaiCongViec: string;
  tenNhomChiTietLoai: string;
  tenChiTietLoai: string;
  tenNguoiTao: string;
  avatar: string;
}

export default JobModel;